Final Commit - Day 1
This exercise is complete and ready for review. I changed the home page file name to index.html, added a web form with a variety of widgets, as well as style rules for the web form, and linked the rb_forms.css sheet to index.html as well as adding rb_formsubmit.js to index.html as well.